Remove unused `skip_if` fn, move utils
Removed `skip_if` function, which is not being used now that there are
subclasses.

Moves `utils.py` back to `cf/utils.py`
